Added block for "delete player" and "resume game"

+++ src/AppMenu/PlayersMenu/Player.js
@@ -2,11 +2,13 @@ import React from "react";
 
 const Player = (props) => {
     const confirmForm = () => {
-        if (window.confirm(`Подтвердить удаление игрока ${props.name}`)) {
-            props.deletePlayer(props.id);
-            props.showStats(true)
-        } else
-            return false
+        if (props.numberOfRound < 8 && props.delPlayerRound === 0) {
+            if (window.confirm(`Подтвердить удаление игрока ${props.name}`)) {
+                props.deletePlayer(props.id);
+                props.showStats(true);
+            }
+        }
+        props.setDelPlayerRound(1)
     };
 
     return (
